requirements.txt
selenium
python-dateutil
ftfy
pytz
pytest